Resolve sonar issues (#19)
* Add const qualifier to irioDrv_t parameters when necessary

* Add branch to sonar action

* Fix linter problems caused by adding const to some irioDrv_t parameters

* Fix missing " in projectBaseDir in sonar action

* Add branch name to sonar action

* Fin incorrect parameter in sonar action

* Remove branch aprameters from sonar action

* Resolve some issues reported by sonar

* Resolve more sonar issues

* Fix Makefile shell if expressions

* Refactor printInfoError to reduce nested statements

* Refactor NISysCfg to replace void* with more meaningul types

* Refactor NISysCfg to prevent copy construction and assignment

* Resolve more sonar issues

* Resolve more sonar issues

* Replace strlen for std::string.size()

+++ src/main/c++/irioCoreCpp/terminals/impl/terminalsAnalogImpl.cpp
@@ -10,13 +10,13 @@ TerminalsAnalogImpl::TerminalsAnalogImpl(ParserManager *parserManager,
 		const NiFpga_Session &session, const Platform &platform) :
 		TerminalsBaseImpl(session) {
 	// Find AI
-	for(size_t i = 0; i < platform.maxAI; ++i) {
+	for(uint32_t i = 0; i < platform.maxAI; ++i) {
 		parserManager->findRegisterEnumAddress(TERMINAL_AI, i,
 				GroupResource::AI, &m_mapAI, true);
 	}
 
 	// Find AO and AOEnable
-	for(size_t i = 0; i < platform.maxAO; ++i) {
+	for(uint32_t i = 0; i < platform.maxAO; ++i) {
 		parserManager->findRegisterEnumAddress(
 			TERMINAL_AO, i, GroupResource::AO, &m_mapAO, true);
 		parserManager->findRegisterEnumAddress(
